+++ CHANGELOG.md
@@ -5,10 +5,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ## [Unreleased]
 
+### Added
+- example config mcc-ulw-sensor
+
+## [2023.8.0]
+
+### Added
+- new daemon: mcc-ulw-sensor
+
 ## [2022.8.0]
 
 ### Added
 - initial release
 
-[Unreleased]: https://github.com/yaq-project/yaqd-mcc/-/compare/v2022.8.0...main
+[Unreleased]: https://github.com/yaq-project/yaqd-mcc/-/compare/v2023.8.0...main
+[2023.8.0]: https://github.com/yaq-project/yaqd-mcc/-/compare/v2022.8.0...v2023.8.0
 [2022.8.0]: https://github.com/yaq-project/yaqd-mcc/-/tags/v2022.8.0